Implement S3 service interface and integration for weather data uploads
