+++ packages/cron_jobs/src/application_business_rules/services_interface/mod.rs
@@ -1,0 +1,2 @@
+pub mod s3_service;
+pub use crate::application_business_rules::services_interface::s3_service::IS3Service;
