+++ packages/cron_jobs/src/application_business_rules/services_interface/s3_service.rs
@@ -1,0 +1,10 @@
+use anyhow::Result;
+use std::future::Future;
+use std::pin::Pin;
+
+pub trait IS3Service {
+    fn upload_to_s3(
+        &self,
+        csv_file_path: String,
+    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send>>;
+}
