+++ packages/cron_jobs/src/interface_adapters/services/mod.rs
@@ -1,0 +1,1 @@
+pub mod s3_service;
