+++ packages/cron_jobs/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod application_business_rules {
+    pub mod services_interface;
     pub mod usecase;
 }
 
@@ -6,4 +7,8 @@ pub mod enterprise_business_rules {
     pub mod domain {
         pub mod model;
     }
+}
+
+pub mod interface_adapters {
+    pub mod services;
 }
